bc/backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
pillow==11.2.1
python-multipart==0.0.20
pydantic==2.11.7
sqlalchemy==2.0.41
aiosqlite==0.21.0
python-dotenv==1.1.1
requests==2.32.3
httpx==0.28.1
diffusers
torch
torchvision
transformers
accelerate
